Move enum map key test and alias cycle test
