+++ integ-tests/typescript/tests/type-aliases.test.ts
@@ -98,4 +98,24 @@ describe('Type aliases tests', () => {
     expect(res).toEqual(data)
     expect(res.json.object.list).toEqual([1, 2, 3])
   })
+
+  it('json type alias as class dependency', async () => {
+    const data = {
+      number: 1,
+      string: 'test',
+      bool: true,
+      list: [1, 2, 3],
+      object: { number: 1, string: 'test', bool: true, list: [1, 2, 3] },
+      json: {
+        number: 1,
+        string: 'test',
+        bool: true,
+        list: [1, 2, 3],
+        object: { number: 1, string: 'test', bool: true, list: [1, 2, 3] },
+      },
+    }
+    const res = await b.TakeRecAliasDep({ value: data })
+    expect(res.value).toEqual(data)
+    expect(res.value.json?.object?.list).toEqual([1, 2, 3])
+  })
 })
